moved latest prompts to configs folder
