+++ src/codeas/use_cases/testing.py
@@ -2,12 +2,12 @@ from typing import List
 
 from pydantic import BaseModel
 
+from codeas.configs import prompts
 from codeas.core.agent import Agent
 from codeas.core.llm import LLMClient
 from codeas.core.metadata import RepoMetadata
 from codeas.core.repo import Repo
 from codeas.core.retriever import ContextRetriever
-from codeas.use_cases import prompts
 
 
 class TestingStep(BaseModel):
